Move battle types to battle folder
