+++ src/battle/components/BattleForm/BattleForm.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { MemoryRouter } from "react-router";
 import BattleForm from "./BattleForm";
-import type { BattleFormData } from "../../../types";
+import type { BattleFormData } from "../../types";
 import { mallorcaBattle } from "../../fixtures";
 
 describe("Given the BattleForm component", () => {
