+++ src/battle/hooks/types.ts
@@ -1,4 +1,4 @@
-import type { BattleFormDataDto, BattlesInfo } from "../../types";
+import type { BattleFormDataDto, BattlesInfo } from "../types";
 import type { BattleDto } from "../dto/types";
 
 export interface UseBattlesStructure {
