+++ src/battle/slice/battlesInfoSlice.tsx
@@ -1,5 +1,5 @@
 import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
-import type { Battle, BattlesInfo } from "../../types";
+import type { Battle, BattlesInfo } from "../types";
 import type { BattlesInfoState } from "./types";
 
 const initialState: BattlesInfoState = {
