+++ src/battle/slice/types.ts
@@ -1,4 +1,4 @@
-import type { BattlesInfo } from "../../types";
+import type { BattlesInfo } from "../types";
 
 export interface BattlesInfoState {
   battlesInfo: BattlesInfo;
