+++ src/battle/types.ts
